fix(packages/monitor): tighten detail panes and unify config rendering
OperatorSessionDetail:
- Fold the standalone HTTP filter card into the Session card as another
  row, with `describeFilter` returning "no filter" when none is set.
  Cuts a near-empty card on most sessions.
- Switch the layout grid from fixed `grid-cols-2` (clamped at 5xl) to
  `grid-cols-1 md:grid-cols-2 xl:grid-cols-3` (clamped at 7xl). On wide
  screens the Session card and Locked ports sit side-by-side and the
  config no longer hogs the full width.

SessionDetail (local sessions):
- Wrap the EventStream block in a `Widget` titled "Events" so it sits
  in the same widget grid as the others (Config, Identity, Ports). No
  more bespoke flat-card treatment for the live event view.
- Default the Config widget to open (matching OperatorSessionDetail);
  user can still collapse it.
- Match the same responsive grid as the operator detail.

ConfigTab:
- Use the new shared `JsonHighlight` (keys / strings / numbers /
  literals) instead of the unstyled `@metalbear/ui` Code component.

JsonHighlight:
- Extracted into its own component so both ConfigTab and the operator
  detail's config widget share a single tokenizer.
